modified: - refactor Premade.Filters.* to implement Filter interface
