+++ src/com/Git/Repository/Generics/Premade/Filters/CHeaderFilesOnly.java
@@ -1,14 +1,13 @@
-package com.Git.Repository.Generics.Premade.Filter;
+package com.Git.Repository.Generics.Premade.Filters;
 
 import com.Git.Repository.Asset;
 import com.Git.Repository.Generics.Filtering.Filter;
-import com.Git.Repository.Generics.Filtering.ListFilter;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CHeaderFilesOnly implements ListFilter<Asset<File>>
+public class CHeaderFilesOnly implements Filter<Asset<File>>
 {
     @Override
     public List<Asset<File>> accept(List<Asset<File>> input)
@@ -33,5 +32,25 @@ public class CHeaderFilesOnly implements ListFilter<Asset<File>>
         }
 
         return newFiles;
+    }
+
+    @Override
+    public boolean accept(Asset<File> file) {
+        int indexOfDot = file.getAsset().getName().indexOf(".");
+        if(!file.getAsset().isDirectory() && indexOfDot != -1)
+        {
+            String extension = file.getAsset().getName().substring(indexOfDot + 1,
+                    file
+                            .getAsset()
+                            .getName()
+                            .length());
+
+            System.out.println(extension);
+
+            if(extension.equals("h"))
+                return true;
+        }
+
+        return false;
     }
 }
